feat(articles): Add user logged in check before searching start
