+++ app/src/main/java/com/makentoshe/habrachan/di/main/articles/ArticlesFlowFragmentModule.kt
@@ -3,14 +3,14 @@ package com.makentoshe.habrachan.di.main.articles
 import com.makentoshe.habrachan.common.database.HabrDatabase
 import com.makentoshe.habrachan.common.database.SessionDao
 import com.makentoshe.habrachan.di.common.ApplicationScope
-import com.makentoshe.habrachan.model.main.articles.ArticlesSearchBroadcastReceiver
+import com.makentoshe.habrachan.model.main.articles.*
 import com.makentoshe.habrachan.view.main.articles.ArticlesFlowFragment
 import toothpick.Toothpick
 import toothpick.config.Module
 import toothpick.ktp.binding.bind
 import toothpick.ktp.delegate.inject
 
-class ArticlesFlowFragmentModule(fragment: ArticlesFlowFragment): Module() {
+class ArticlesFlowFragmentModule(fragment: ArticlesFlowFragment) : Module() {
 
     private val database by inject<HabrDatabase>()
 
@@ -20,5 +20,14 @@ class ArticlesFlowFragmentModule(fragment: ArticlesFlowFragment): Module() {
 
         val searchBroadcastReceiver = ArticlesSearchBroadcastReceiver()
         bind<ArticlesSearchBroadcastReceiver>().toInstance(searchBroadcastReceiver)
+        bind<ArticlesSearchEpoxyController>().toInstance(getArticlesSearchEpoxyController())
     }
+
+    private fun getArticlesSearchEpoxyController() = ArticlesSearchEpoxyController(
+        database.session(),
+        ArticlesSearchTopEpoxyModel.Factory(),
+        ArticlesSearchAllEpoxyModel.Factory(),
+        ArticlesSearchInterestingEpoxyModel.Factory(),
+        ArticlesSearchSubscriptionEpoxyModel.Factory()
+    )
 }
